+++ src/services/venv.service.ts
@@ -1,7 +1,7 @@
 import { err, ok, Result } from "neverthrow";
 import apiService from "./api.service";
 import type { AxiosError } from "axios";
-import type { RLGymLearnApiExceptionModel, VenvConfig } from "rlgym-learn-client";
+import type { PackageInfo, RLGymLearnApiExceptionModel, VenvConfig } from "rlgym-learn-client";
 
 class VenvService {
     async createEnvironment(projectId: string, pythonExecutable: string): Promise<Result<VenvConfig, AxiosError<RLGymLearnApiExceptionModel>>>{
@@ -13,6 +13,12 @@ class VenvService {
 
     async deleteEnvironment(projectId: string): Promise<Result<string, AxiosError<RLGymLearnApiExceptionModel>>>{
         return apiService.venvApi.deleteVenv(projectId).then(
+            (r) => ok(r.data)
+        ).catch((r) => err(r))
+    }
+
+    async getPackages(projectId: string): Promise<Result<Record<string, PackageInfo>, AxiosError<RLGymLearnApiExceptionModel>>>{
+        return apiService.venvApi.listPackages(projectId).then(
             (r) => ok(r.data)
         ).catch((r) => err(r))
     }